split data_access from the pipeline module
